fix qiantangzhengxin bug

+++ geaflow/geaflow-dsl/geaflow-dsl-connector/geaflow-dsl-connector-odps/src/main/java/com/antgroup/geaflow/dsl/connector/odps/utils/OdpsConnectorUtils.java
@@ -41,7 +41,7 @@ public class OdpsConnectorUtils {
             case INT:
                 return itype instanceof LongType || itype instanceof IntegerType;
             case BIGINT:
-                return itype instanceof IntegerType;
+                return itype instanceof LongType;
             case FLOAT:
                 return itype instanceof FloatType || itype instanceof DoubleType;
             case DOUBLE:
